+++ Cargo.toml
@@ -18,9 +18,12 @@ tokio = { version = "1.45.1", features = ["full"] }
 rustemon = { version = "4.2.0" }
 dirs = "6.0.0"
 futures = "0.3.31"
-reqwest = { version = "0.12.19", features = ["json"] }
+reqwest = { version = "0.12.20", features = ["json"] }
 serde = { version = "1.0.219", features = ["serde_derive"] }
 serde_json = "1.0.140"
+memmap2 = "0.9.5"
+dashmap = { version = "6.1.0", features = ["serde"] }
+anyhow = "1.0.98"
 
 [dependencies.i18n-embed]
 version = "0.15.4"
