+++ src/config.rs
@@ -1,18 +1,29 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 use cosmic::{
-    cosmic_config::{self, cosmic_config_derive::CosmicConfigEntry, CosmicConfigEntry},
+    cosmic_config::{self, CosmicConfigEntry, cosmic_config_derive::CosmicConfigEntry},
     theme,
 };
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Default, Clone, CosmicConfigEntry, Eq, PartialEq)]
+#[derive(Debug, Clone, CosmicConfigEntry, Eq, PartialEq)]
 #[version = 1]
 pub struct Config {
     pub app_theme: AppTheme,
     pub first_run_completed: bool,
     pub pokemon_per_row: usize,
     pub type_filtering_mode: TypeFilteringMode,
+}
+
+impl Default for Config {
+    fn default() -> Self {
+        Self {
+            app_theme: Default::default(),
+            first_run_completed: false,
+            pokemon_per_row: 3,
+            type_filtering_mode: Default::default(),
+        }
+    }
 }
 
 #[derive(Clone, Copy, Debug, Default, Eq, PartialEq, Serialize, Deserialize)]
