+++ src/entities.rs
@@ -1,0 +1,40 @@
+// SPDX-License-Identifier: GPL-3.0-only
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct StarryPokemon {
+    pub pokemon: StarryPokemonData,
+    pub sprite_path: Option<String>,
+    pub encounter_info: Option<Vec<StarryPokemonEncounterInfo>>,
+}
+
+/// Core Pokémon data
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct StarryPokemonData {
+    pub id: i64,
+    pub name: String,
+    pub weight: i64,
+    pub height: i64,
+    pub types: Vec<String>,
+    pub abilities: Vec<String>,
+    pub stats: StarryPokemonStats,
+}
+
+/// Pokémon statistics
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct StarryPokemonStats {
+    pub hp: i64,
+    pub attack: i64,
+    pub defense: i64,
+    pub sp_attack: i64,
+    pub sp_defense: i64,
+    pub speed: i64,
+}
+
+/// Pokémon encounter information
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct StarryPokemonEncounterInfo {
+    pub city: String,
+    pub games_method: Vec<String>,
+}
