+++ src/utils.rs
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
-use std::fs;
+use std::fmt::Display;
 
-use crate::app::StarryPokemonStats;
+use anyhow::Error;
 
-const APP_ID: &str = "dev.mariinkys.StarryDex";
+use crate::entities::StarryPokemonStats;
 
 pub fn capitalize_string(input: &str) -> String {
     let words: Vec<&str> = input.split('-').collect();
@@ -22,10 +22,6 @@ pub fn capitalize_string(input: &str) -> String {
         .collect();
 
     capitalized_words.join(" ")
-}
-
-pub fn scale_numbers(num: i64) -> f64 {
-    (num as f64) / 10.0
 }
 
 pub fn parse_pokemon_stats(stats: &[rustemon::model::pokemon::PokemonStat]) -> StarryPokemonStats {
@@ -53,56 +49,47 @@ pub fn parse_pokemon_stats(stats: &[rustemon::model::pokemon::PokemonStat]) -> S
     starry_stats
 }
 
-pub async fn download_image(
-    client: &reqwest::Client,
-    image_url: String,
-    pokemon_name: String,
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    let resources_path = dirs::data_dir()
-        .unwrap()
-        .join(APP_ID)
-        .join("resources")
-        .join("sprites");
-
-    if !resources_path.exists() {
-        fs::create_dir_all(&resources_path).expect("Failed to create the resources path");
-    }
-
-    let image_filename = format!("{}_front.png", pokemon_name);
-    let image_path = resources_path.join(&pokemon_name).join(&image_filename);
-
-    // Check if file already exists
-    if tokio::fs::metadata(&image_path).await.is_ok() {
-        return Ok(());
-    }
-
-    let response = client.get(&image_url).send().await?;
-    if response.status().is_success() {
-        let bytes = response.bytes().await?;
-        let path = std::path::PathBuf::from(&image_path);
-        if let Some(parent) = path.parent() {
-            tokio::fs::create_dir_all(parent).await?;
-        }
-        tokio::fs::write(&image_path, &bytes).await?;
-        Ok(())
-    } else {
-        Err(Box::new(std::io::Error::other(format!(
-            "Failed to download image. Status: {}",
-            response.status()
-        ))))
-    }
+pub fn scale_numbers(num: i64) -> f64 {
+    (num as f64) / 10.0
 }
 
 pub fn remove_dir_contents<P: AsRef<std::path::Path>>(path: P) -> std::io::Result<()> {
-    for entry in fs::read_dir(path)? {
+    for entry in std::fs::read_dir(path)? {
         let entry = entry?;
         let path = entry.path();
 
         if path.is_dir() {
-            fs::remove_dir_all(path)?;
+            std::fs::remove_dir_all(path)?;
         } else {
-            fs::remove_file(path)?;
+            std::fs::remove_file(path)?;
         }
     }
     Ok(())
+}
+
+#[derive(Debug)]
+pub enum StarryError {
+    Generic(Error),
+    NoCacheFound,
+}
+
+impl Display for StarryError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match &self {
+            StarryError::Generic(error) => write!(f, "{}", error),
+            StarryError::NoCacheFound => write!(f, "No cache found!"),
+        }
+    }
+}
+
+impl From<std::io::Error> for StarryError {
+    fn from(value: std::io::Error) -> Self {
+        Self::Generic(value.into())
+    }
+}
+
+impl From<serde_json::Error> for StarryError {
+    fn from(value: serde_json::Error) -> Self {
+        Self::Generic(value.into())
+    }
 }
